+++ test/dimensions/correlationdim.jl
@@ -1,9 +1,8 @@
 using ChaosTools, DynamicalSystemsBase
 using Test
-using StatsBase
 using Statistics
 using ChaosTools.DynamicalSystemsBase
-using ChaosTools.DelayEmbeddings: standardize
+using ChaosTools.DelayEmbeddings
 
 test_value = (val, vmin, vmax) -> @test vmin <= val <= vmax
 
@@ -12,12 +11,12 @@ println("\nTesting correlation sum...")
     @testset "Henon Map" begin
         ds = Systems.henon()
         ts = trajectory(ds, 5000)
-        R = maximum(maxima(ts) - minima(ts))
+        R = 1.5maximum(maxima(ts) - minima(ts))
         # check for right normalisation with various combinations of `q` and `w`
-        @test correlationsum(ts, R+1; q = 2, w = 0) ≈ 1
-        @test correlationsum(ts, R+1; q = 2, w = 1) ≈ 1
-        @test correlationsum(ts, R+1; q = 2.2, w = 0) ≈ 1
-        @test correlationsum(ts, R+1; q = 2.2, w = 1) ≈ 1
+        @test correlationsum(ts, R; q = 2, w = 0) ≈ 1
+        @test correlationsum(ts, R; q = 2, w = 1) ≈ 1
+        @test correlationsum(ts, R; q = 2.2, w = 0) ≈ 1
+        @test correlationsum(ts, R; q = 2.2, w = 1) ≈ 1
         es = 10 .^ range(-3, stop = 0, length = 7)
         cs1 = correlationsum(ts, es)
         dim1 = linear_region(log.(es), log.(cs1))[2]
@@ -41,8 +40,8 @@ println("\nTesting correlation sum...")
     end
 end
 
-println("\nTesting correlation sum with boxing beforehand...")
-@testset "Theilers correlation boxing algorithm" begin
+println("\nTesting box-assisted correlation sum...")
+@testset "Box-assisted correlation sum" begin
     @testset "Henon Map" begin
         ds = Systems.henon()
         ts = standardize(trajectory(ds, 10000))
@@ -71,7 +70,7 @@ end
 
 
 println("\nTesting the fixed mass correlation sum...")
-@testset "Fixed mass correlation sum by Grassberger" begin
+@testset "Fixed mass correlation sum" begin
     @testset "Henon Map" begin
         ds = Systems.henon()
         ts = trajectory(ds, 10000)